Update merge_lanes.R
Added recursive=TRUE as un-merged fastq-files downloaded from BaseSpace ends up in subfolders in the original in path, so user don't have to move them into the same folder prior to running merging functions.  
Also added a more refined way to find unique names, as the previous one did not work if length of names were different. This one uses 4 lanes as default, but with the option for the user to change number of lanes (unsure how common 2 lane sequencing is?)