Update merge_lanes.R
Makes code easier to work with files downloaded with Windows version of BaseSpace Downloader.
Additionally - removed the $ in the regex for the ending of file name, as this does not work when file names differ in length. The addition as is now ("_L00.*") is not bulletproof (if a sample would be named something with _L00"), but at least works with the standard file names in Illumina.